requirements.txt
Django==3.1.2
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
kolo==1.0.4
